Fix partial construction of the graph
If a vertex is not connected to *any* zettel (regardless of direction)
it is excluded from the graph, because neuron used `LAM.edges` only to
construct it (which builds the graph from edges, not vertexes). The
solution is to construct two graphs -- one built from edges, and another
from vertices -- and then combine them.

Fixes #32