ZettelStore: more helpful lookup failure message
I made a typo and transposed a couple of numbers in a zettel link. I was
getting an error message like `No such zettel: ZettelDateID 2020-03-19`,
but since there were a bunch of links in the doc, it took me a while to
figure out that this corresponded to the `[2011401]` link in my doc.
I've changed the error message for date IDs to include the text as it
would actually appear in the link in addition to the parsed day/index.

It would probably also be nice to get source line numbers on error
messages, but that sounds like a more ambitious effort.

I've never really written Haskell, so if I'm doing anything weird, it's
for no good reason, and please let me know.