Change neuron-search to not be nix-templated
This makes this package easier to built from hackage or in nixpkgs where
we cannot use IFD.

With nix the existence of the dependencies for the back-script should be
reasonable save because we just have to track them in the default.nix
and in nixpkgs.

This change has still the problem that some one trying to build neuron
from hackage will need make sure that src-bash is in their PATH before
compiling.
Maybe we want to add that to the README?