Make neuron new title argument optional
Now, if the user does not pass an argument for the title, a default '# Zettel created on <date>' is used.
Made the change in the action that writes the file instead of in the command line parser because that is written with an applicative parser so I couldn't figure out if I could use the value of a parameter (date) to build the default for another parameter (title).