Avoid requiring `builtins.currentSystem`
This commit modifies the `default.nix` and `shell.nix` files to take
arguments with defaults which are provided to `project.nix` (which no
longer takes arguments). This is necessary because on systems which use
Nix flakes, `builtins.currentSystem` cannot be used and there was no
mechanism for users to stop Neuron from using `currentSystem` when
importing it.

Signed-off-by: David Wood <david@davidtw.co>