Disable use of reflex-platform
This change migrates the default.nix back to a configuration that does
not use reflex-platform.

The default.nix now tries to be similar to how neuron would be build in
nixpkgs.

The pinned nixpkgs version is now pinned to a very recent version of
nixos unstable.

The default ghc on nixpkgs right now is 8.8.3, so this change also
includes some changes to make compilation possible with that ghc and the
newest package versions from nixpkgs.