Define searchOptions and enable query expansion
This appears to be what is required to get the term "ins" to begin
returning results for "install" etc. ("in" is a stop word, so does not
trigger the search).

Searching for just "ins" returns more than just "install" results,
because it could expand to other words. It appears that it also expands
to "integration", which makes enough sense to me since that could be a
typo. "integration" drops out of the results once the query term is
expanded to "inst".

This also highlights a potentially better and simpler way to handle
tag-search--using the field boost/suppression in the search options to
limit tag search to only the tag field.